requirements.txt
psutil
requests
python-dotenv
rich
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
